chore: Chat UI
Built a customer wrapper for the clickable buttons using emotion so buttons appear over chat - had to build custom as Twilio Paste Box only allows z-index of 90 and this does not bring content to "the front" of the view.
Tried to get a resizing of chat window on presence of clickable buttons, it shrinks when present but does not resize when no clickables as required (this code is commented out in the useEffect).